Update styles.css
Background style targeted to the body selector.
justify-content:center & align-items:center included in the background to center the child element (.wrapper). Thanks for the tip @RibeiroPorto!
height: 100vh added to represent 100% of the viewport's height, or the full height of the screen. 
Overflow: hidden included in the wrapper section to hide text overflow when scrolling.